services/metadata_extractor/requirements.txt
pika==1.3.2
minio==7.2.7
psycopg2-binary==2.9.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
